Fix merge regression bugs relating to transaction notifications (#328)
There were two major regression related bugs from merging in the new
database. The first was failure to properly notify transactions in the
stake tree. The second was that the wrong sentinel value was being
used to terminate the rescanning loop, which normally causes
registration for relevant outpoints and addresses. Both bugs are
patched.