deps: Update to latest commits. (#362)
This updates the glide.yaml to remove the incorrect seelog version so it
can make use of the latest commit and adds the test import for testify
used by the dcrec package.  Finally it uses glide update to update the
glide.lock file accordingly.

All commits in between the old and newest commits in the lock file have
been reviewed.