mining/cpuminer: modify config to add new GetMiningAddr func
This commit modifies the existing `cpuminer.Config` struct to replace
with the `MiningAddrs` attribute and manual generation address
selection with a new function closure. When called, this function
closure returns a mining address to be used within the next generated
block. This modification adds a layer of abstraction over the storage
and selection logic for obtaining a new mining address.