mempool: Optimize the votes map slices.
This optimizes the vote map by making better use of Go's slice properties.
In particular, instead of creating a slice with empty entries and
iterating to find the next empty entry, just create the slice with a cap
and append entries as needed.  This allows the code simply take the len
of the slice and range over it without needing additional code to deal
with nil entries.

While here, it also makes a few comments and returns more consistent
with the rest of the code base.