mempool: Rename GetVoteHashesForBlock & remove err.
This renames the GetVoteHashesForBlock on the mempool type to
VoteHashesForBlock so it is named according to Effective Go guidelines
as required by the project's code contribution guidelines and removes
the error return and paths from the function.

The error has been removed since the function should not be assuming how
the caller wishes to use the data.  Consequently, now when there are no
votes for a given block hash, the function simply returns a nil slice to
reflect that.  The caller can then decide if the lack of vote hashes for
a block is an error or not.

Also, the error path regarding an unset vote hash in the metadata is an
impossible condition because the mempool has full control over the vote
insertion and thus will never insert vote metadata without it.