chaingen: Add package for generating test chains.
This introduces a new package named chaingen that is useful for
generating full-blown valid test chains.  It is primarily a refactor of
the generator code that was previously in fullblocktests.

It also updates the fullblocktests package to make use of this new
chaingen package and removes the stakeversiontests package in favor of
performing all of the same tests inline making use of the new chaingen
package.