dcrctl: Remove --terminal feature (#580)
The problem of securely prompting for secrets and not leaving secrets in shell
history can be solved in better ways, such as implementing a tool to read a
secret and write it back to standard output, allowing the secret to be piped to
dcrctl.  Luckily dcrctl already supports this feature!

No (official) tool exists so far, but the workflow would look like this:

  promptsecret | dcrctl --wallet walletpassphrase - 60

While here removing unneeded deps from the glide files, also throw in an
unrelated dcrrpcclient dep update.