rpcserver: Don't return irrelevant cooked values.
The calculated values in the getvoteinfo command are essentially
meaningless when a vote has not been started.  The argument for having
them was to gauge some sort of progress but that was a flawed argument.
We have decided to move the raw data into the gtstakeversions call where
it belongs.