chaincfg: strictly enforce agendas assumptions.
Agendas require some complicated and not always obvious assumptions.  In
order to prevent bad agendas from making it into the code we validate
them at init time and panic if an agenda violates assumptions.  This
will not adversely impact the user experience but will force developers
to properly design dcrd agendas.

These are the required tests to ensure vote integrity:
* All choices shall be consecutive and enumerated.
* Mask only contains consecutive 1 bits set.
* Ensure IsIgnore and IsNo are never both set to true.
* Ensure there is one and only one IsNo choice set to true.
* Ensure there is one and only one IsIgnore choice set to true.
* Ensure that the abstain choice is 0.
* Ensure that the number of choices does not exceed the mask bits
  permissible size.
* Ensure that choice bits match the mask bits.
* Ensure there are no duplicate choice ids.
* Ensure there are no duplicate vote ids.