chaincfg: Allow non-std transactions on testnet2.
The test network should allow any scripts and transaction types that are
valid according to consensus rules as opposed to the more strict
standardness rules that apply to mainnet.  This allows new script forms
and types to be tested before they go to mainnet.

Also, this is the same setting as the previous test network.