rpcserver/chain: Bounds check getstakeversions.
This adds bounds check for the RPC parameters for the getstakeversions
command to both the handler and the chain function itself which provides
the data.

Also, while here simplify the for loop in GetStakeVersions to check
prevNode in the for loops condition instead of in a separate if block,
return more appropriate error types for RPC parameter type errors, and
use a naked type cast of the command to be consistent with all of the
other handlers.