rpcserver: Return handler errors to RPC client.
Only the first parameter to the rpcInternalError function is returned to
RPC clients, so any errors that originate in the handlers need to be
passed as the first parameter.

The second context parameter is to allow additional context to be added
to errors coming from underlying calls.  There is no additional context
for errors generated by the RPC handlers themselves.