indexers: Minimize differences with upstream code.
This modifies all files in the indexers package to minimize the
differences between them and the upstream code in order to facilitate
easier syncs due to less merge conflicts as a result of superfluous
changes.

While here, it also corrects a few comments that were not updated
properly for the changes in Decred.