chain/indexers: Allow reorg of block one.
This modifies the code to permit a reorganization that includes block
one.  Even though this will never happen on the main network since it's
buried so deep, it is entirely possible for the first block to be
reorged away on test chains.

Not only is it more technically accurate to permit a reorg of the first
block, so long as the new one satisfies all of consensus rules, it also
facilitates upcoming test code which relies on that capability.

It should be noted that this is technically a hard fork, however, it
only applies to the first block and so it would now require a reorg of
over 150,000 blocks which is completely infeasible.