mempool: Update comments for dust calcs.
This updates the comments in the isDust function to match the current
default relay fee of 1e5, correct tx size values, and make them
more consistent with the field order.  It's nice to have the most recent
values as a reference in the comments.

Also, while here, add a couple of tests just below and above the dust
point for the current default relay fee.