fullblocktests: Add more consensus tests.
This adds several new tests to the consensus tests based upon their
upstream equivalents as well as some Decred-specific cases and updates
the chaingen package accordingly.

The following is a summary of the new tests:

* Negative proof-of-work limit (bits header field)
* Spend from transaction that does not exist
* Max size block with canonical encoding which exceeds max size with
  non-canonical encoding
* Too many signature operations with offending sigop before invalid data
  push
* Max and too many signature operations via pay-to-script-hash redeem
  scripts
* Invalid opcode in dead execution path
* Reorganization of txns with OP_RETURN outputs
* Spend of an OP_RETURN output
* Transaction with multiple OP_RETURN outputs
* Large max-sized block reorganization test (disabled by default since
  it takes a long time and a lot of memory to run)

It also modifies a few tests to make them a little less reliant on the
specific parameters chosen such as replacing the number of votes with
the expected tickets per block divided by 2 instead of just subtracting
3 which assumes there are specifically 5 votes.

Finally, it adds some comments regarding some upstream tests that were
not included due to the specific conditions they test not applying to
Decred.