config: Improve config file handling.
This improves the default config file creation process so that it will
only create a default config file when the user has not overridden the
config file path.

It also now works properly when a different appdata path is provided in
that it will still create a default config file in that new path if one
doesn't already exist, but only if the user did not also override the
config file path.

Finally, instead of copying a separate sample config file from the
filesystem, which didn't work properly under several scenarios such as
when it wasn't already available or the appdata directory was changed,
the code now specifies the entire sample config as a string in the
source code.  This means it will work as intended under all scenarios.