blockchain: Load all block headers into the block index on init.
Currently only the blocks in the active chain are loaded into the
block index on initialization. This instead iterates over the entire
block index bucket in LevelDB and loads all nodes.

In order to do so efficiently, we switch the key format for the blocks
in LevelDB to be prefixed with the block height encoded as big-endian,
so when iterating over keys, we always see blocks in order of height.